sql: Add system.jobs table and migration to add it to existing DBs
Unfortunately, this required playing around a bit with the table
creation code to get the permissions and ID right on the new table.
This feels a little dangerous, so let me know what you think. I
effectively ripped it straight out of #7073.

The alternative was to add the table directly to kv, bypassing the SQL
layer, but that seems dangerous to do anytime other than initial
bootstrapping and won't work great for future changes like adding a
new column or index to a system table.