gossip: Enforce 0 <= nodeSet.len() <= nodeSet.maxSize
Before this change, we'd let nodes open up more than maxSize outgoing
connections because new connections would only be added to the outgoing
nodeSet after they had been established (so that we'd have the peerID).
Until a connection was established and we received the peer's ID, the
calls to outgoing.hasSpace in gossip.go would let more connections be
established.

In small clusters, this isn't a big deal, but in large clusters it can
cause far too many connections to get opened, leading to a lot of
connections getting bounced around as they get refused by nodes that are
at their limit of incoming connections. Making matters even worse is
that doDisconnected doesn't check whether there's room to connect to
the forwardAddr, meaning that once a cluster has
`total outgoing conns` > `total allowed incoming conns`, those extra
connections can just keep bouncing around indefinitely.

By ensuring we don't open too many connections, we can keep that from
happening.

On a 100 cycle simulation of a 64 node cluster, this takes the number of
refused connections down from ~3700 to ~750 (or from ~5300 to ~875) if
you take out the previous change to avoid duplicates in tightenNetwork).

More notably, the number of refused connections stops growing completely
shortly after ~50 cycles, indicating a lack of the indefinite thrashing
that we've seen on such large clusters in the wild.