allocsim: Adjust latency's role in lease transfer calculations
Make it play a smaller role in low-latency situations, and a more
dominant role in high latency situations. This is a slightly risky
change. While reducing the effect in low-latency situations I expect
will almost always be a good thing, the effect in high-latency
situations might be for the worse sometimes if request loads aren't
steady.