[WIP] experimental cRPC protocol
The cRPC protocol is built on top of http2 using only http2 data
frames. The core read and write loops are similar to gRPC, but trimmed
down to their essence. A ton of stuff is missing before this could
actually be used, such as context cancellation propagation, trace
propagation, chunking of large requests/responses into frames, and
proper error handling. This PR was done as a proof of concept to see
what performance impact replacing gRPC with something else could
achieve.

On a single-node cluster with the local server optimization disabled,
gRPC on a read-mostly workload shows:

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)
   10.0s        0         111904        11189.4      1.4      1.2      3.0      5.8     18.9

cRPC:

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)
   10.0s        0         183923        18389.3      0.9      0.7      1.8      3.7     62.9

Unfortunately, the effects on a multi-node cluster are less dramatic.

I think it would be possible to make cRPC wire compatible with
gRPC. This would require sending a headers frame with every RPC, but the
headers are small (7 bytes) and constant for this use case. We'd also
have to send a settings frame as the first frame on the connection, but
that is trivial.

See #17370.